Initiation of login added to the project
Post and get are in place.
Small implementation of a redirect in order to validate the user is added.  If the password is correct, the user will be redirected to a special page, otherwise if the password is not correct the user will be redirected back to the login page.